Message Stats refactored into component
