+++ components/ChromaMeep.js
@@ -4,8 +4,8 @@ import { useState } from "react";
 import Image from "next/image";
 
 export default function ChromaMeep() {
-    //Tilt for possible pause feature?
-    const [tilting, setTilt] = useState(true);
+//  Tilt for possible pause feature?
+//  const [tilting, setTilt] = useState(true);
     const [boom, setBoom] = useState(false);
 
     const handleClick = () => {
@@ -15,11 +15,11 @@ export default function ChromaMeep() {
             setBoom(false);
         }, 50);
 
-        setTilt((a) => !a);
-
-        setTimeout(() => {
-            setTilt(true);
-        }, 10);
+//        setTilt((a) => !a);
+//
+//        setTimeout(() => {
+//            setTilt(true);
+//       }, 10);
     };
 
     return (
